backend/pyproject.toml
[project]
name = "recipe-app-backend"
version = "0.5.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",

  "SQLAlchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",

  "python-multipart>=0.0.9",

  "python-jose[cryptography]>=3.3",
  "passlib>=1.7.4",

  "email-validator>=2.1",

  "reportlab>=4.0",

  "itsdangerous>=2.2.0",
  "sqladmin==0.16.1",
  "psycopg2-binary==2.9.9",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]

[tool.ruff]
line-length = 100
